All the graphs are done for the eda section are done and have been standardized. Still not able to keep only 2 boxes for the combined histograms of the numerical variables (age, avg_glucose_level, bmi). Graphs not shown on the eda section (because of irrelevance|non significance) were put on the Appendix section. Some verifications need to be done for the titles of graphs. The preprocessed section graphs were not done yet, I will start now. No Analysis were done yet on the balanced & stratifies analysis part (will start tonight, it should not be long).
All the graphs are done for the eda section are done and have been standardized. Still not able to keep only 2 boxes for the combined histograms of the numerical variables (age, avg_glucose_level, bmi). Graphs not shown on the eda section (because of irrelevance|non significance) were put on the Appendix section. Some verifications need to be done for the titles of graphs. The preprocessed section graphs were not done yet, I will start now. No Analysis were done yet on the balanced & stratifies analysis part (will start tonight, it should not be long).